+++ src/app/dataset-view/additional-components/metadata-component/components/set-transform/components/engine-section/engine-section.types.ts
@@ -2,4 +2,5 @@ export enum Engine {
     Spark = "Spark",
     Flink = "Flink",
     DataFusion = "DataFusion",
+    RisingWave = "RisingWave",
 }